Bug fix for trimming inputs, data formatting, clear logs button, and other improvements

+++ src/main/java/com/mcasaje/simplejavaweather/models/weather/SimpleWeatherDataFactory.java
@@ -1,12 +1,14 @@
 package com.mcasaje.simplejavaweather.models.weather;
 
+import java.math.BigDecimal;
+
 public class SimpleWeatherDataFactory implements WeatherDataFactory {
 
     @Override
     public SimpleWeatherData create(String locationName,
                                     String weatherDescription,
-                                    Double tempInCelsius,
-                                    Double tempInFahrenheit,
+                                    BigDecimal tempInCelsius,
+                                    BigDecimal tempInFahrenheit,
                                     String sunriseStandardTime,
                                     String sunsetStandardTime) {
         return new SimpleWeatherData(locationName,
